+++ elroy/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.0.66"
+__version__ = "0.0.68"
 
 from importlib_resources import files