+++ elroy/tools/function_caller.py
@@ -188,10 +188,6 @@ def get_json_type(py_type: Type) -> Union[str, Dict[str, Any]]:
     raise ValueError(f"Unsupported type: {py_type}")
 
 
-def get_modules():
-    return []
-
-
 ERROR_PREFIX = "**Tool call resulted in error: **"
 
 
@@ -250,7 +246,8 @@ def get_function_schema(function: Callable) -> Dict:
 
     assert function.__doc__ is not None, f"Function {function.__name__} has no docstring"
     parsed_docstring = parse(function.__doc__)
-    description = parsed_docstring.short_description or parsed_docstring.long_description
+    assert parsed_docstring.description, f"Function {function.__name__} has no description in docstring"
+    description = parsed_docstring.description.strip()
     docstring_dict = {p.arg_name: p.description for p in parse(function.__doc__).params}
 
     signature = inspect.signature(function)