+++ tests/debugging.py
@@ -1,6 +1,0 @@
-# Debugging utils, should not be used in live code.
-
-
-def do_debug(x):
-    print(x)
-    return x